# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Traccar",
    "version": "6.1",
    "description": "Traccar GPS tracking server API documentation. To use the API you need to have a server instance. For testing purposes you can use one of free [demo servers](https://www.traccar.org/demo-server/). For production use you can install your own server or get a [subscription service](https://www.traccar.org/product/tracking-server/).",
    "contact": {
      "name": "FOX GPS Support",
      "url": "https://www.foxgps.com.br/",
      "email": "sac@foxgps.com.br"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "servers": [
    {
      "url": "https://api.foxgps.com.br/api",
      "description": "FOX GPS Server"
    }
  ],
  "security": [
    {
      "basicAuth": []
    }
  ],
...
  "paths": {
...
    "/permissions": {
      "post": {
        "summary": "Link an Object to another Object",
        "operationId": "postObject",
        "tags": [
          "Permissions"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Permission"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content",
            "content": {}
          },
          "400": {
            "description": "No permission",
            "content": {}
          }
        },
        "x-codegen-request-body-name": "body"
      },
      "delete": {
        "summary": "Unlink an Object from another Object",
        "operationId": "deleteObject",
        "tags": [
          "Permissions"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Permission"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content",
            "content": {}
          }
        },
        "x-codegen-request-body-name": "body"
      }
    },
...
    "/notifications": {
      "get": {
        "summary": "Fetch a list of Notifications",
        "description": "Without params, it returns a list of Notifications the user has access to",
        "tags": [
          "Notifications"
        ],
        "parameters": [
          {
            "name": "all",
            "in": "query",
            "description": "Can only be used by admins or managers to fetch all entities",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "userId",
            "in": "query",
            "description": "Standard users can use this only with their own _userId_",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "deviceId",
            "in": "query",
            "description": "Standard users can use this only with _deviceId_s, they have access to",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "groupId",
            "in": "query",
            "description": "Standard users can use this only with _groupId_s, they have access to",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "refresh",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Notification"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a Notification",
        "tags": [
          "Notifications"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Notification"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Notification"
                }
              }
            }
          }
        },
        "x-codegen-request-body-name": "body"
      }
    },
...
    "/notifications/test": {
      "post": {
        "summary": "Send test notification to current user via Email and SMS",
        "tags": [
          "Notifications"
        ],
        "responses": {
          "204": {
            "description": "Successful sending",
            "content": {}
          },
          "400": {
            "description": "Could happen if sending has failed",
            "content": {}
          }
        }
      }
    },
...
    "/attributes/computed": {
      "get": {
        "summary": "Fetch a list of Attributes",
        "description": "Without params, it returns a list of Attributes the user has access to",
        "tags": [
          "Attributes"
        ],
        "parameters": [
          {
            "name": "all",
            "in": "query",
            "description": "Can only be used by admins or managers to fetch all entities",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "userId",
            "in": "query",
            "description": "Standard users can use this only with their own _userId_",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "deviceId",
            "in": "query",
            "description": "Standard users can use this only with _deviceId_s, they have access to",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "groupId",
            "in": "query",
            "description": "Standard users can use this only with _groupId_s, they have access to",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "refresh",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Attribute"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create an Attribute",
        "tags": [
          "Attributes"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Attribute"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Attribute"
                }
              }
            }
          }
        },
        "x-codegen-request-body-name": "body"
      }
    },
    "/attributes/computed/{id}": {
      "put": {
        "summary": "Update an Attribute",
        "tags": [
          "Attributes"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Attribute"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Attribute"
                }
              }
            }
          }
        },
        "x-codegen-request-body-name": "body"
      },
      "delete": {
        "summary": "Delete an Attribute",
        "tags": [
          "Attributes"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content",
            "content": {}
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Permission": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "integer",
            "description": "User id, can be only first parameter"
          },
          "deviceId": {
            "type": "integer",
            "description": "Device id, can be first parameter or second only in combination with userId"
          },
          "groupId": {
            "type": "integer",
            "description": "Group id, can be first parameter or second only in combination with userId"
          },
          "geofenceId": {
            "type": "integer",
            "description": "Geofence id, can be second parameter only"
          },
          "notificationId": {
            "type": "integer",
            "description": "Notification id, can be second parameter only"
          },
          "calendarId": {
            "type": "integer",
            "description": "Calendar id, can be second parameter only and only in combination with userId"
          },
          "attributeId": {
            "type": "integer",
            "description": "Computed attribute id, can be second parameter only"
          },
          "driverId": {
            "type": "integer",
            "description": "Driver id, can be second parameter only"
          },
          "managedUserId": {
            "type": "integer",
            "description": "User id, can be second parameter only and only in combination with userId"
          },
          "commandId": {
            "type": "integer",
            "description": "Saved command id, can be second parameter only"
          }
        },
        "description": "This is a permission map that contain two object indexes. It is used to link/unlink objects. Order is important. Example: { deviceId:8, geofenceId: 16 }"
      },
...
      "Notification": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "type": {
            "type": "string"
          },
          "always": {
            "type": "boolean"
          },
          "web": {
            "type": "boolean"
          },
          "mail": {
            "type": "boolean"
          },
          "sms": {
            "type": "boolean"
          },
          "calendarId": {
            "type": "integer"
          },
          "attributes": {
            "type": "object",
            "properties": {}
          }
        }
      },
...
      "Attribute": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "description": {
            "type": "string"
          },
          "attribute": {
            "type": "string"
          },
          "expression": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "description": "String|Number|Boolean"
          }
        }
      },
...
    },
...
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "description": "Basic HTTP authorization with _email_ and _password_",
        "scheme": "basic"
      }
    }
  }
}